Enviar mensajes por plantillas

+++ index.php
@@ -1,6 +1,8 @@
 <?php
 
 use Twilio\Rest\Content\V1\ContentInstance;
+use Twilio\Rest\Messaging\V1\ServiceInstance;
+
 
 require __DIR__ . '/vendor/autoload.php';
 
@@ -26,11 +28,11 @@ class Twilio2
     }
 
     /**
-     * listarPlatilla
-     *
+     * listarPlatillaContentAndApprovals
+     * https://www.twilio.com/docs/content/content-api-resources#fetch-content-and-approvals
      * @return Respuesta
      */
-    function listarPlatilla(): Respuesta
+    function listarPlatillaContentAndApprovals(): Respuesta
     {
 
         $url = "{$this->urlBase}v1/Content";
@@ -46,7 +48,7 @@ class Twilio2
         // Realiza la solicitud GET
         $response   = curl_exec($ch);
         $http_code  = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-        $curl_error     = curl_error($ch);
+        $curl_error = curl_error($ch);
 
         // Cierra la conexión cURL
 
@@ -55,11 +57,11 @@ class Twilio2
     }
 
     /**
-     * listarPlatillaAprovados
+     * listarPlatilla
      * https://www.twilio.com/docs/content/content-api-resources#fetch-mapping-between-legacy-wa-and-content-templates
-     * @return void
+     * @return ContentAndApprovalsInstance
      */
-    function listarPlatillaAprovados()
+    function listarPlatilla()
     {
 
         $contentAndApprovals = $this->twilio->content->v1->contentAndApprovals
@@ -75,15 +77,67 @@ class Twilio2
         // @property array|null $types
         // @property array|null $approvalRequests
 
-        foreach ($contentAndApprovals as $record) {
-            print("SID: {$record->sid}");
-            echo "<br>";
-            print("friendlyName: {$record->friendlyName}");
-            echo "<br>";
-            print("accountSid: {$record->accountSid}");
-            echo "<br>";
-        }
+        // foreach ($contentAndApprovals as $record) {
+        // echo "<br>";
+        // print("SID: {$record->sid}");
+        // echo "<br>";
+        // print("friendlyName: {$record->friendlyName}");
+        // echo "<br>";
+        // print("accountSid: {$record->accountSid}");
+        // echo "<br>";
+        // print("approvalRequests:");
+        // echo  json_encode($record->approvalRequests);
+        // echo "<br>";
+        // print("status:");
+        // echo  json_decode(json_encode($record->approvalRequests))->status;
+        // print("Significado status:");
+        // echo "<strong>" . Plantilla::ESTADO[json_decode(json_encode($record->approvalRequests))->status] . "</strong>";
+        // echo "<br>";
+        // print("rejection_reason:");
+        // echo  json_decode(json_encode($record->approvalRequests))->rejection_reason;
+        // echo "<br>";
+        // }
         return  $contentAndApprovals;
+    }
+
+    function platillasAprobadas()
+    {
+        $aprobados = [];
+
+        foreach ($this->listarPlatilla() as $record) {
+
+            if ("approved" ==  json_decode(json_encode($record->approvalRequests))->status) {
+                $aprobados[] = $record;
+
+                // * El valor record->sid es el que se debe usar para mandar las plantillas en los whatsapp en el contentSid Ejmplo = "contentSid" => "HX9f6291...",
+                print("SID: {$record->sid}");
+                echo "<br>";
+                print("friendlyName: {$record->friendlyName}");
+                echo "<br>";
+                echo "<strong>" . Plantilla::ESTADO[json_decode(json_encode($record->approvalRequests))->status] . "</strong>";
+                echo "<br>";
+                echo "<br>";
+                echo json_encode($record->types);
+                foreach ((($record->types)) as $key => $value) {
+                    echo  $key;
+                    echo "<br>";
+                    if (key_exists("body", $value)) {
+
+                        echo $value['body'];
+                        echo "<br>";
+                    }
+
+                    if (key_exists("media", $value)) {
+
+                        echo $value['media'][0];
+                        echo "<br>";
+                    }
+                }
+                echo "<br>";
+            }
+        }
+
+        return $aprobados;
     }
 
     /**
@@ -134,6 +188,47 @@ class Twilio2
 
         return new Respuesta($http_code, $response, $curl_error);
     }
+
+    /**
+     * messaging_service_sid
+     * Son los Servicios que nos permiten enviar Mensajes (Sender)
+     * @return ServiceInstance
+     */
+    function messaging_service_sid()
+    {
+
+        $services = $this->twilio->messaging->v1->services->read();
+
+        // Itera a través de los servicios y muestra sus SIDs
+
+        // @property string|null $sid
+
+        // @property string|null $accountSid
+
+        // @property string|null $friendlyName
+
+        // @property \DateTime|null $dateCreated
+
+        // @property \DateTime|null $dateUpdated
+
+        // @property string|null $inboundRequestUrl
+
+        // @property string|null $inboundMethod
+        // * El  $service->sid se usa para mandar mensajes en el FROM ejemplo =  "from"       => "MG1bf093cd...",
+        foreach ($services as $service) {
+            echo "Service SID: " . $service->sid . "\n";
+            echo "<br>";
+            echo "Service friendlyName: " . $service->friendlyName . "\n";
+            echo "<br>";
+            echo "Service inboundRequestUrl: " . $service->inboundRequestUrl . "\n";
+            echo "<br>";
+            echo "Service inboundMethod: " . $service->inboundMethod . "\n";
+            echo "<br>";
+            echo "<br>";
+        }
+
+        return $services;
+    }
 }
 
 class Respuesta
@@ -182,6 +277,17 @@ class Plantilla implements Convertible
     private $language;
     private $variables = [];
     private $types = [];
+
+    const ESTADO = [
+        'unsubmitted'   => "Indica que la plantilla no se ha enviado a Twilio o WhatsApp para ningún tipo de aprobación. Estas plantillas aún se pueden usar en sesión para algunos canales y en algunas sesiones de WA sujetas a los requisitos de aprobación de WhatsApp enumerados anteriormente.",
+        'received'      => "Indica que Twilio ha recibido la solicitud de aprobación de la plantilla. Todavía no está en revisión por WhatsApp.",
+        'pending'       => "Indica que la plantilla está siendo revisada por WhatsApp. La revisión puede tardar hasta 24 horas.",
+        'approved'      => "La plantilla fue aprobada por WhatsApp y se puede utilizar para notificar a los clientes.",
+        'rejected'      => "La plantilla ha sido rechazada por WhatsApp durante el proceso de revisión.",
+        'paused'        => "WhatsApp ha pausado la plantilla debido a los comentarios negativos recurrentes de los usuarios finales, que generalmente resultan de las acciones de 'bloquear' y 'reportar spam' asociadas con la plantilla. Las plantillas de mensajes con este estado no se pueden enviar a los usuarios finales.",
+        'disabled'      => "WhatsApp ha desactivado la plantilla debido a comentarios negativos recurrentes de los usuarios finales o por violar una o más de las políticas de WhatsApp. Las plantillas de mensajes con este estado no se pueden enviar a los usuarios finales.",
+
+    ];
 
     public function __construct($friendly_name, $language, $variables, Types $types)
     {
@@ -341,16 +447,54 @@ $sid        = "";
 $token      = "";
 
 $twilio2    = new Twilio2($sid, $token);
+$services = $twilio2->messaging_service_sid();
 
+$twilio2->platillasAprobadas();
 
-// $twilio2->listarPlatillaAprovados();
-echo "<br>";
-$contenteId     = "";
-echo $twilio2->obtener($contenteId)->sid;
 // echo "<br>";
-// echo $twilio2->listarPlatilla()->response;
+// echo "<br>";
+// $contenteId     = "HX7...";
+// echo json_encode($twilio2->obtener($contenteId)->types);
+// echo "<br>";
+// echo "<br>";
+// $contenteId     = "HX...";
+// echo json_encode($twilio2->obtener($contenteId)->types);
 
-$plantilla = new Plantilla('Prueba 2', 'en', [], new Types('twilio/text', "Hola Mundo. Twilio2"));
+
+
+
+$client = new Twilio\Rest\Client($sid, $token);
+
+
+
+// foreach ($services as $service) {
+//     $message = $client->messages
+//         ->create(
+//             "whatsapp:+", // to
+//             [
+//                 "from"       =>  $service->sid,
+//                 "contentSid" => "HX...",
+//             ]
+//         );
+//     echo "<br>";
+//     print($message->sid);
+// }
+echo "<br>";
+$message = $client->messages
+    //   ->create("whatsapp:+", // to
+    ->create(
+        "whatsapp:+", // to
+        [
+            "from"       => "MG...",
+            "contentSid" => "HX..",
+        ]
+    );
+echo "<br>";
+print($message->sid);
+
+
+
+// $plantilla = new Plantilla('Prueba 2', 'en', [], new Types('twilio/text', "Hola Mundo. Twilio2"));
 
 // echo "<br>";
 // echo $plantilla->toJsonApi();
@@ -360,7 +504,7 @@ $plantilla = new Plantilla('Prueba 2', 'en', [], new Types('twilio/text', "Hola 
 // $respuesta = $twilio2->crearPlantilla($plantilla);
 // echo $respuesta->toJson();
 // echo "<br>";
-echo $twilio2->listarPlatilla()->response;
-echo "<br>";
-$twilio2->listarPlatillaAprovados();
-echo "<br>";
+// echo $twilio2->listarPlatilla()->response;
+// echo "<br>";
+// $twilio2->listarPlatilla();
+// echo "<br>";
